// src/model/activity.cfg.ts
/*
 *  __________________
 *  Qaobee
 *  __________________
 *
 *  Copyright (c) 2015.
 *  Qaobee
 *  All Rights Reserved.
 *
 *  NOTICE: All information contained here is, and remains
 *  the property of Qaobee and its suppliers,
 *  if any. The intellectual and technical concepts contained
 *  here are proprietary to Qaobee and its suppliers and may
 *  be covered by U.S. and Foreign Patents, patents in process,
 *  and are protected by trade secret or copyright law.
 *  Dissemination of this information or reproduction of this material
 *  is strictly forbidden unless prior written permission is obtained
 *  from Qaobee.
 */

export class ActivityCfg {
    periodDuration: number;
    nbMaxPlayers: number;
    nbMinPlayers: number;
    nbPeriod: number;
    nbTimeout: number;
    timeoutDuration: number;
    yellowCardMax: number;
    exclusionTempo: number;
    halfTimeDuration: number;
}